requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.77
CacheControl==0.13.1
cachetools==5.3.2
cement==2.8.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
cloudinary==1.36.0
colorama==0.4.3
colorthief==0.2.1
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.7
django-channels-jwt-auth-middleware==1.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fcm-django==1.0.12
firebase-admin==5.4.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
haversine==2.8.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.7
pathspec==0.10.1
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
semantic-version==2.8.5
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.1.9
whitenoise==6.6.0
zope.interface==6.1
